requirements.txt
fastapi[all]==0.110.0
uvicorn[standard]==0.29.0
sqlalchemy[asyncio]==2.0.29
geoalchemy2==0.14.6
asyncpg==0.29.0
pydantic==2.7.0
pydantic-settings==2.2.1
alembic==1.13.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
redis[hiredis]==5.0.3
shapely==2.0.3
psycopg2-binary==2.9.9
greenlet==3.0.3
